moving enums to their own files and inside a sub folder
