+++ Assets/Mirage/Runtime/Transport/SocketLayer/Enums.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6ef9a95bf405d3145a8f0c3eb1ef8269
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
