+++ Assets/Mirage/Runtime/Transport/SocketLayer/Enums/PacketType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4e58d122f3debcb4ebb5315885a997c3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
