+++ Assets/Mirage/Runtime/Transport/SocketLayer/Enums/RejectReason.cs
@@ -1,0 +1,16 @@
+﻿namespace Mirage.SocketLayer
+{
+    /// <summary>
+    /// Reason for reject sent from server
+    /// </summary>
+    public enum RejectReason
+    {
+        None = 0,
+        ServerFull = 1,
+        Timeout = 2,
+        /// <summary>
+        /// Closed called locally before connect
+        /// </summary>
+        ClosedByPeer = 3,
+    }
+}
