+++ Assets/Mirage/Runtime/Transport/SocketLayer/Enums/RejectReason.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4fc031488fac9c84694f3c20449b35c7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
